+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**
 
 - Improve frame drop detection algorithm. ([#304](https://github.com/getsentry/vroom/pull/304))
+- Accept and return the time a profile started at in a timestamp field on Android. ([#306](https://github.com/getsentry/vroom/pull/306))
 
 **Internal**